fix: Add missing uint256 conversion for totalSupply() in test
- Fix type mismatch in testFuzz_Transfer_Invariant
- totalSupply() returns Balance type, needs uint256() conversion for assertEq
- This was causing compilation error in the fuzz test

+++ contracts/test/Token.t.sol
@@ -276,8 +276,8 @@ contract TokenTest is Test {
 
         assertEq(
             token.balanceOf(alice) + token.balanceOf(bob),
-            token.totalSupply()
+            uint256(token.totalSupply())
         );
-        assertEq(token.totalSupply(), mintAmount);
+        assertEq(uint256(token.totalSupply()), mintAmount);
     }
 }
